refactor: 설명 수정 변경된 UI에 따른 리팩토링
